pyproject.toml
[project]
name = "pyturb"
version = "0.1.0"
description = "Ocean turbulent microstructure processing"
readme = "README.md"
authors = [
    { name = "Jesse Cusack", email = "jesse.cusack@oregonstate.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.0",
    "netcdf4>=1.7.3",
    "numpy>=2.3.2",
    "profinder>=0.2.4",
    "scipy>=1.16.1",
    "typer-slim>=0.20.0",
    "xarray>=2025.10.1",
]

[project.scripts]
pyturb = "pyturb.cli:cli"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "bokeh>=3.8.0",
    "ipykernel>=7.0.0",
    "matplotlib>=3.10.7",
    "mypy>=1.17.1",
    "pip>=25.2",
    "ruff>=0.12.11",
    "scipy-stubs>=1.16.1.1",
    "tqdm>=4.67.1",
]
